get rid of redux
