+++ demos/automerge-repo-todos/package.json
@@ -27,14 +27,13 @@
     "@localfirst/auth-shared": "workspace:*",
     "@localfirst/auth-syncserver": "workspace:*",
     "@localfirst/crypto": "workspace:*",
-    "@reduxjs/toolkit": "^1.9.2",
+    "@uidotdev/usehooks": "^2.4.1",
     "autoprefixer": "^10.4.13",
     "classnames": "^2.3.2",
     "clipboard": "^2.0.11",
     "postcss": "^8.4.21",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-redux": "^8.0.5",
     "ua-parser-js": "^1.0.37"
   },
   "devDependencies": {
@@ -42,7 +41,6 @@
     "@types/node": "^18.18.13",
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.9",
-    "@types/react-redux": "^7.1.32",
     "@types/ua-parser-js": "^0.7.39",
     "@vitejs/plugin-react": "^4.2.1",
     "tailwindcss": "^3.3.7",
