+++ demos/automerge-repo-todos/src/components/SignOut.tsx
@@ -1,12 +1,7 @@
-import { useDispatch } from 'react-redux'
-import { actions } from '../store/reducer'
-const { logout } = actions
+import { useLocalState } from '../hooks/useLocalState'
 
 export const SignOut = () => {
-  const dispatch = useDispatch()
-  const signOut = () => {
-    dispatch(logout())
-  }
+  const { signOut } = useLocalState()
   return (
     <button className="button button-primary button-sm" onClick={signOut}>
       Sign out
