+++ demos/automerge-repo-todos/src/hooks/useRootDocument.ts
@@ -1,12 +1,11 @@
 import { stringifyAutomergeUrl, type AutomergeUrl } from '@automerge/automerge-repo'
 import { useDocument } from '@automerge/automerge-repo-react-hooks'
 import { assert } from '@localfirst/auth-shared'
-import { useSelector } from 'react-redux'
-import { selectRootDocumentId } from '../store/selectors'
 import { type SharedState } from '../types'
+import { useLocalState } from './useLocalState'
 
 export const useRootDocument = () => {
-  const rootDocumentId = useSelector(selectRootDocumentId)
+  const { rootDocumentId } = useLocalState()
   assert(rootDocumentId)
   const rootDocumentUrl: AutomergeUrl = stringifyAutomergeUrl({
     documentId: rootDocumentId,
