--- demos/automerge-repo-todos/src/store/initialState.ts
@@ -1,3 +1,0 @@
-import type { LocalState } from '../types'
-
-export const initialState: LocalState = {}
